requirements.txt
# ============================================================
# Technology Pipeline Tracker — Python dependencies
# ============================================================
# Install with:  pip install -r requirements.txt
# Python 3.11+  recommended
# ============================================================

# ── Web framework & server ───────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# ── Anthropic / Claude API ───────────────────────────────────
anthropic==0.34.2

# ── HTTP clients ─────────────────────────────────────────────
requests==2.32.3
httpx==0.27.2          # used by anthropic SDK internally

# ── Reddit ───────────────────────────────────────────────────
praw==7.8.1
prawcore==2.4.0

# ── News ─────────────────────────────────────────────────────
newsapi-python==0.2.7

# ── Academic APIs (optional helpers; some tools use HTTP directly) ────────
arxiv==2.4.1
semanticscholar==0.12.0

# ── Google Trends ────────────────────────────────────────────
trendspyg==0.4.2

# ── HTML parsing / scraping ──────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.3.0            # faster BS4 parser
playwright==1.58.0     # run: playwright install  (downloads browser binaries)

# ── Vector embeddings + RAG ──────────────────────────────────
chromadb==0.5.15
sentence-transformers==3.1.1

# ── Databases ────────────────────────────────────────────────
pymongo==4.10.1
neo4j==5.24.0

# ── Environment / config ─────────────────────────────────────
python-dotenv==1.0.1

# ── Data utilities ───────────────────────────────────────────
python-dateutil==2.9.0.post0   # flexible date parsing

# ── Typing (stdlib backport for 3.9/3.10 compat) ────────────
# typing-extensions is pulled in by pydantic/anthropic

# ── Dev / optional ───────────────────────────────────────────
# ipython            # for interactive exploration
# pytest             # for running tests
# pytest-asyncio     # for async FastAPI tests
